+++ src/utils/Ability/getAbility.js
@@ -3,6 +3,11 @@ const  ABILITY_URL = 'https://pokeapi.co/api/v2/ability';
 
 
 //=======================$ability helper functions====================//
+async function getAbilityData(message) {
+    const ability = message.content.split(" ")[1];
+    const abData = await getAbility(ability);
+    return abData;
+}
 
 async function getAbility(ability){
     let response = await fetch(`${ABILITY_URL}/${ability}`);
@@ -10,4 +15,4 @@ async function getAbility(ability){
 }
 
 //=======================EXPORTS====================//
-module.exports= { getAbility };
+module.exports= { getAbility, getAbilityData };
