+++ src/utils/Type/getType.js
